--- a/src/transformers/models/fsmt/modeling_fsmt.py
+++ b/src/transformers/models/fsmt/modeling_fsmt.py
@@ -341,8 +341,9 @@ class SinusoidalPositionalEmbedding(nn.Embedding):
     def __init__(self, num_positions: int, embedding_dim: int, padding_idx: int):
         self.make_weight(num_positions, embedding_dim, padding_idx)
 
-    def make_weight(self, num_positions, embedding_dim, padding_idx):
+    def make_weight(self, num_positions, embedding_dim, padding_idx, dtype=torch.float32):
         weight = self.get_embedding(num_positions, embedding_dim, padding_idx)
+        weight = weight.to(dtype)
         if not hasattr(self, "weight"):
             # in ___init__
             super().__init__(num_positions, embedding_dim, padding_idx, _weight=weight)
@@ -419,7 +420,7 @@ class SinusoidalPositionalEmbedding(nn.Embedding):
         max_pos = self.padding_idx + 1 + seq_len + past_key_values_length
         if max_pos > self.weight.size(0):
             # expand embeddings if needed
-            self.make_weight(max_pos, self.embedding_dim, self.padding_idx)
+            self.make_weight(max_pos, self.embedding_dim, self.padding_idx, self.weight.dtype)
 
         return super().forward(positions)