@stas00 2021-05-15
@patil-suraj, what needs to be done to complete this? 

Last we talked there was a performance regression and the suggestion was to test Bart's performance pre and post its original refactoring.